+++ BankProject/BankProject.Data/Entities/LoanApplication.cs
@@ -8,7 +8,7 @@ public class LoanApplication : IEntity<Guid>
     public Guid Id { get; set; }
     public LoanType LoanType { get; set; }
     
-    public float LoanAmount { get; set; }
+    public decimal LoanAmount { get; set; }
     
     public int LoanTerm { get; set; }
